Fix and document the use of %*<N><datatype> in binary formats to skip unwanted fields

+++ src/datafile.h
@@ -1,5 +1,5 @@
 /*
- * $Id: datafile.h,v 1.25 2008/03/13 19:53:27 sfeam Exp $
+ * $Id: datafile.h,v 1.26 2008/03/14 02:56:24 sfeam Exp $
  */
 
 /* GNUPLOT - datafile.h */
@@ -183,7 +183,7 @@ typedef struct df_binary_type_struct {
 } df_binary_type_struct;
 
 typedef struct df_column_bininfo_struct {
-    short skip_bytes;
+    long skip_bytes;
     df_binary_type_struct column;
 } df_column_bininfo_struct;
